backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
mcp>=1.0.0
httpx>=0.26.0
numpy-financial>=1.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
scipy>=1.11.0
